+++ app/(auth)/layout.tsx
@@ -5,13 +5,15 @@ import "../../styles/globals.css";
 
 export default function AuthLayout({ children }: { children: ReactNode }) {
   return (
-    <>
-      <Header />
-      <main className="container mx-auto p-4">
-        <h1>Welcome to auth Section</h1>
-        {children}
-      </main>
-      <Footer />
-    </>
+    <html lang="en">
+      <body>
+        <Header />
+        <main className="container mx-auto p-4">
+          <h1>Welcome to auth Section</h1>
+          {children}
+        </main>
+        <Footer />
+      </body>
+    </html>
   );
 }
